app/requirements.txt
click==7.1.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
Werkzeug==1.0.1
requests==2.27.1
